requirements.txt
gradio
langchain
langchain-community
langchain-groq
huggingface-hub
sentence-transformers
transformers
torchgradio
langchain
langchain-community
langchain-groq
huggingface-hub
sentence-transformers
transformers
torch
python-dotenv

